requirements.txt
pydantic
requests
python-dotenv
matplotlib
numpy